requirements.txt
streamlit
google-generativeai
pandas
python-dotenv
requests
reportlab
langchain
langchain-community
pypdf
langchain-google-genai
